[v1 channel] move dynamic filters to its own BUILD target
