+++ BUILD
@@ -4010,7 +4010,6 @@ grpc_cc_library(
         "//src/core:client_channel/client_channel_factory.cc",
         "//src/core:client_channel/client_channel_filter.cc",
         "//src/core:client_channel/client_channel_plugin.cc",
-        "//src/core:client_channel/dynamic_filters.cc",
         "//src/core:client_channel/global_subchannel_pool.cc",
         "//src/core:client_channel/load_balanced_call_destination.cc",
         "//src/core:client_channel/local_subchannel_pool.cc",
@@ -4025,7 +4024,6 @@ grpc_cc_library(
         "//src/core:client_channel/client_channel.h",
         "//src/core:client_channel/client_channel_factory.h",
         "//src/core:client_channel/client_channel_filter.h",
-        "//src/core:client_channel/dynamic_filters.h",
         "//src/core:client_channel/global_subchannel_pool.h",
         "//src/core:client_channel/load_balanced_call_destination.h",
         "//src/core:client_channel/local_subchannel_pool.h",
@@ -4100,6 +4098,7 @@ grpc_cc_library(
         "//src/core:construct_destruct",
         "//src/core:context",
         "//src/core:dual_ref_counted",
+        "//src/core:dynamic_filters",
         "//src/core:error",
         "//src/core:error_utils",
         "//src/core:exec_ctx_wakeup_scheduler",
