+++ build_autogenerated.yaml
@@ -234,7 +234,6 @@ libs:
   - src/core/client_channel/config_selector.h
   - src/core/client_channel/connector.h
   - src/core/client_channel/direct_channel.h
-  - src/core/client_channel/dynamic_filters.h
   - src/core/client_channel/global_subchannel_pool.h
   - src/core/client_channel/lb_metadata.h
   - src/core/client_channel/load_balanced_call_destination.h
@@ -932,6 +931,7 @@ libs:
   - src/core/lib/channel/channel_stack_builder.h
   - src/core/lib/channel/channel_stack_builder_impl.h
   - src/core/lib/channel/connected_channel.h
+  - src/core/lib/channel/dynamic_filters.h
   - src/core/lib/channel/promise_based_filter.h
   - src/core/lib/compression/compression_internal.h
   - src/core/lib/compression/message_compress.h
@@ -1410,7 +1410,6 @@ libs:
   - src/core/client_channel/client_channel_plugin.cc
   - src/core/client_channel/client_channel_service_config.cc
   - src/core/client_channel/direct_channel.cc
-  - src/core/client_channel/dynamic_filters.cc
   - src/core/client_channel/global_subchannel_pool.cc
   - src/core/client_channel/lb_metadata.cc
   - src/core/client_channel/load_balanced_call_destination.cc
@@ -1917,6 +1916,7 @@ libs:
   - src/core/lib/channel/channel_stack_builder.cc
   - src/core/lib/channel/channel_stack_builder_impl.cc
   - src/core/lib/channel/connected_channel.cc
+  - src/core/lib/channel/dynamic_filters.cc
   - src/core/lib/channel/promise_based_filter.cc
   - src/core/lib/compression/compression.cc
   - src/core/lib/compression/compression_internal.cc
@@ -2495,7 +2495,6 @@ libs:
   - src/core/client_channel/config_selector.h
   - src/core/client_channel/connector.h
   - src/core/client_channel/direct_channel.h
-  - src/core/client_channel/dynamic_filters.h
   - src/core/client_channel/global_subchannel_pool.h
   - src/core/client_channel/lb_metadata.h
   - src/core/client_channel/load_balanced_call_destination.h
@@ -2675,6 +2674,7 @@ libs:
   - src/core/lib/channel/channel_stack_builder.h
   - src/core/lib/channel/channel_stack_builder_impl.h
   - src/core/lib/channel/connected_channel.h
+  - src/core/lib/channel/dynamic_filters.h
   - src/core/lib/channel/promise_based_filter.h
   - src/core/lib/compression/compression_internal.h
   - src/core/lib/compression/message_compress.h
@@ -3058,7 +3058,6 @@ libs:
   - src/core/client_channel/client_channel_plugin.cc
   - src/core/client_channel/client_channel_service_config.cc
   - src/core/client_channel/direct_channel.cc
-  - src/core/client_channel/dynamic_filters.cc
   - src/core/client_channel/global_subchannel_pool.cc
   - src/core/client_channel/lb_metadata.cc
   - src/core/client_channel/load_balanced_call_destination.cc
@@ -3203,6 +3202,7 @@ libs:
   - src/core/lib/channel/channel_stack_builder.cc
   - src/core/lib/channel/channel_stack_builder_impl.cc
   - src/core/lib/channel/connected_channel.cc
+  - src/core/lib/channel/dynamic_filters.cc
   - src/core/lib/channel/promise_based_filter.cc
   - src/core/lib/compression/compression.cc
   - src/core/lib/compression/compression_internal.cc
