+++ gRPC-C++.podspec
@@ -322,7 +322,6 @@ Pod::Spec.new do |s|
                       'src/core/client_channel/config_selector.h',
                       'src/core/client_channel/connector.h',
                       'src/core/client_channel/direct_channel.h',
-                      'src/core/client_channel/dynamic_filters.h',
                       'src/core/client_channel/global_subchannel_pool.h',
                       'src/core/client_channel/lb_metadata.h',
                       'src/core/client_channel/load_balanced_call_destination.h',
@@ -1022,6 +1021,7 @@ Pod::Spec.new do |s|
                       'src/core/lib/channel/channel_stack_builder.h',
                       'src/core/lib/channel/channel_stack_builder_impl.h',
                       'src/core/lib/channel/connected_channel.h',
+                      'src/core/lib/channel/dynamic_filters.h',
                       'src/core/lib/channel/promise_based_filter.h',
                       'src/core/lib/compression/compression_internal.h',
                       'src/core/lib/compression/message_compress.h',
@@ -1752,7 +1752,6 @@ Pod::Spec.new do |s|
                               'src/core/client_channel/config_selector.h',
                               'src/core/client_channel/connector.h',
                               'src/core/client_channel/direct_channel.h',
-                              'src/core/client_channel/dynamic_filters.h',
                               'src/core/client_channel/global_subchannel_pool.h',
                               'src/core/client_channel/lb_metadata.h',
                               'src/core/client_channel/load_balanced_call_destination.h',
@@ -2452,6 +2451,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/channel/channel_stack_builder.h',
                               'src/core/lib/channel/channel_stack_builder_impl.h',
                               'src/core/lib/channel/connected_channel.h',
+                              'src/core/lib/channel/dynamic_filters.h',
                               'src/core/lib/channel/promise_based_filter.h',
                               'src/core/lib/compression/compression_internal.h',
                               'src/core/lib/compression/message_compress.h',
