+++ gRPC-Core.podspec
@@ -319,8 +319,6 @@ Pod::Spec.new do |s|
                       'src/core/client_channel/connector.h',
                       'src/core/client_channel/direct_channel.cc',
                       'src/core/client_channel/direct_channel.h',
-                      'src/core/client_channel/dynamic_filters.cc',
-                      'src/core/client_channel/dynamic_filters.h',
                       'src/core/client_channel/global_subchannel_pool.cc',
                       'src/core/client_channel/global_subchannel_pool.h',
                       'src/core/client_channel/lb_metadata.cc',
@@ -1529,6 +1527,8 @@ Pod::Spec.new do |s|
                       'src/core/lib/channel/channel_stack_builder_impl.h',
                       'src/core/lib/channel/connected_channel.cc',
                       'src/core/lib/channel/connected_channel.h',
+                      'src/core/lib/channel/dynamic_filters.cc',
+                      'src/core/lib/channel/dynamic_filters.h',
                       'src/core/lib/channel/promise_based_filter.cc',
                       'src/core/lib/channel/promise_based_filter.h',
                       'src/core/lib/compression/compression.cc',
@@ -2686,7 +2686,6 @@ Pod::Spec.new do |s|
                               'src/core/client_channel/config_selector.h',
                               'src/core/client_channel/connector.h',
                               'src/core/client_channel/direct_channel.h',
-                              'src/core/client_channel/dynamic_filters.h',
                               'src/core/client_channel/global_subchannel_pool.h',
                               'src/core/client_channel/lb_metadata.h',
                               'src/core/client_channel/load_balanced_call_destination.h',
@@ -3386,6 +3385,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/channel/channel_stack_builder.h',
                               'src/core/lib/channel/channel_stack_builder_impl.h',
                               'src/core/lib/channel/connected_channel.h',
+                              'src/core/lib/channel/dynamic_filters.h',
                               'src/core/lib/channel/promise_based_filter.h',
                               'src/core/lib/compression/compression_internal.h',
                               'src/core/lib/compression/message_compress.h',
