+++ grpc.gemspec
@@ -203,8 +203,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/client_channel/connector.h )
   s.files += %w( src/core/client_channel/direct_channel.cc )
   s.files += %w( src/core/client_channel/direct_channel.h )
-  s.files += %w( src/core/client_channel/dynamic_filters.cc )
-  s.files += %w( src/core/client_channel/dynamic_filters.h )
   s.files += %w( src/core/client_channel/global_subchannel_pool.cc )
   s.files += %w( src/core/client_channel/global_subchannel_pool.h )
   s.files += %w( src/core/client_channel/lb_metadata.cc )
@@ -1413,6 +1411,8 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/channel/channel_stack_builder_impl.h )
   s.files += %w( src/core/lib/channel/connected_channel.cc )
   s.files += %w( src/core/lib/channel/connected_channel.h )
+  s.files += %w( src/core/lib/channel/dynamic_filters.cc )
+  s.files += %w( src/core/lib/channel/dynamic_filters.h )
   s.files += %w( src/core/lib/channel/promise_based_filter.cc )
   s.files += %w( src/core/lib/channel/promise_based_filter.h )
   s.files += %w( src/core/lib/compression/compression.cc )
