+++ src/python/grpcio/grpc_core_dependencies.py
@@ -46,7 +46,6 @@ CORE_SOURCE_FILES = [
     'src/core/client_channel/client_channel_plugin.cc',
     'src/core/client_channel/client_channel_service_config.cc',
     'src/core/client_channel/direct_channel.cc',
-    'src/core/client_channel/dynamic_filters.cc',
     'src/core/client_channel/global_subchannel_pool.cc',
     'src/core/client_channel/lb_metadata.cc',
     'src/core/client_channel/load_balanced_call_destination.cc',
@@ -556,6 +555,7 @@ CORE_SOURCE_FILES = [
     'src/core/lib/channel/channel_stack_builder.cc',
     'src/core/lib/channel/channel_stack_builder_impl.cc',
     'src/core/lib/channel/connected_channel.cc',
+    'src/core/lib/channel/dynamic_filters.cc',
     'src/core/lib/channel/promise_based_filter.cc',
     'src/core/lib/compression/compression.cc',
     'src/core/lib/compression/compression_internal.cc',
