+++ tools/doxygen/Doxyfile.core.internal
@@ -973,8 +973,6 @@ src/core/client_channel/config_selector.h \
 src/core/client_channel/connector.h \
 src/core/client_channel/direct_channel.cc \
 src/core/client_channel/direct_channel.h \
-src/core/client_channel/dynamic_filters.cc \
-src/core/client_channel/dynamic_filters.h \
 src/core/client_channel/global_subchannel_pool.cc \
 src/core/client_channel/global_subchannel_pool.h \
 src/core/client_channel/lb_metadata.cc \
@@ -2219,6 +2217,8 @@ src/core/lib/channel/channel_stack_builder_impl.cc \
 src/core/lib/channel/channel_stack_builder_impl.h \
 src/core/lib/channel/connected_channel.cc \
 src/core/lib/channel/connected_channel.h \
+src/core/lib/channel/dynamic_filters.cc \
+src/core/lib/channel/dynamic_filters.h \
 src/core/lib/channel/promise_based_filter.cc \
 src/core/lib/channel/promise_based_filter.h \
 src/core/lib/compression/GEMINI.md \
